+++ src/libs/data/sql.c
@@ -36,15 +36,15 @@ NSP_FUNCTION(libnsp_data_sql_common_escape)
 	l2 = 1;
 	for (; *se; se++) {
 		if (nc_strncmp(se, s2, l2) != 0) continue;
-		nsp_strcat(N, &N->r, ss, se - ss);
+		nsp_strcat(N, &N->r, ss, (unsigned long)(se - ss));
 		nsp_strcat(N, &N->r, "''", 2);
 		ss = se += l2;
 		if (*se) { --se; continue; }
-		nsp_strcat(N, &N->r, ss, se - ss);
+		nsp_strcat(N, &N->r, ss, (unsigned long)(se - ss));
 		break;
 	}
 	if (se > ss) {
-		nsp_strcat(N, &N->r, ss, se - ss);
+		nsp_strcat(N, &N->r, ss, (unsigned long)(se - ss));
 	}
 	return 0;
 #undef __FN__
